Move survey component to folder
